+++ client-app/src/ui-kit/breadcrumb.component.tsx
@@ -1,0 +1,86 @@
+import { Slot } from "@radix-ui/react-slot"
+import { ChevronRight, MoreHorizontal } from "lucide-react"
+import { cn } from "@/helpers/concat-class-names.helper"
+import { forwardRef, ComponentPropsWithoutRef, ReactNode, ComponentProps } from "react"
+
+export const Breadcrumb = forwardRef<
+    HTMLElement,
+    ComponentPropsWithoutRef<"nav"> & {
+        separator?: ReactNode
+    }
+>(({ ...props }, ref) => <nav ref={ref} aria-label="breadcrumb" {...props} />)
+
+Breadcrumb.displayName = "Breadcrumb"
+
+export const BreadcrumbList = forwardRef<HTMLOListElement, ComponentPropsWithoutRef<"ol">>(
+    ({ className, ...props }, ref) => (
+        <ol
+            ref={ref}
+            className={cn(
+                "flex flex-wrap items-center gap-1.5 break-words text-sm text-muted-foreground sm:gap-2.5",
+                className,
+            )}
+            {...props}
+        />
+    ),
+)
+
+BreadcrumbList.displayName = "BreadcrumbList"
+
+export const BreadcrumbItem = forwardRef<HTMLLIElement, ComponentPropsWithoutRef<"li">>(
+    ({ className, ...props }, ref) => (
+        <li ref={ref} className={cn("inline-flex items-center gap-1.5", className)} {...props} />
+    ),
+)
+
+BreadcrumbItem.displayName = "BreadcrumbItem"
+
+export const BreadcrumbLink = forwardRef<
+    HTMLAnchorElement,
+    ComponentPropsWithoutRef<"a"> & {
+        asChild?: boolean
+    }
+>(({ asChild, className, ...props }, ref) => {
+    const Comp = asChild ? Slot : "a"
+
+    return <Comp ref={ref} className={cn("transition-colors hover:text-foreground", className)} {...props} />
+})
+
+BreadcrumbLink.displayName = "BreadcrumbLink"
+
+export const BreadcrumbPage = forwardRef<HTMLSpanElement, ComponentPropsWithoutRef<"span">>(
+    ({ className, ...props }, ref) => (
+        <span
+            ref={ref}
+            role="link"
+            aria-disabled="true"
+            aria-current="page"
+            className={cn("font-normal text-foreground", className)}
+            {...props}
+        />
+    ),
+)
+
+BreadcrumbPage.displayName = "BreadcrumbPage"
+
+export const BreadcrumbSeparator = ({ children, className, ...props }: ComponentProps<"li">) => (
+    <li role="presentation" aria-hidden="true" className={cn("[&>svg]:w-3.5 [&>svg]:h-3.5", className)} {...props}>
+        {children ?? <ChevronRight />}
+    </li>
+)
+
+BreadcrumbSeparator.displayName = "BreadcrumbSeparator"
+
+export const BreadcrumbEllipsis = ({ className, ...props }: ComponentProps<"span">) => (
+    <span
+        role="presentation"
+        aria-hidden="true"
+        className={cn("flex h-9 w-9 items-center justify-center", className)}
+        {...props}
+    >
+        <MoreHorizontal className="h-4 w-4" />
+        <span className="sr-only">More</span>
+    </span>
+)
+
+BreadcrumbEllipsis.displayName = "BreadcrumbElipssis"
