+++ client-app/src/ui-kit/collapsible.component.tsx
@@ -1,0 +1,1 @@
+export { Collapsible, CollapsibleTrigger, CollapsibleContent } from "@radix-ui/react-collapsible"
