+++ client-app/src/ui-kit/input.component.tsx
@@ -1,0 +1,18 @@
+import { cn } from "@/helpers/concat-class-names.helper"
+import { forwardRef, ComponentProps } from "react"
+
+export const Input = forwardRef<HTMLInputElement, ComponentProps<"input">>(({ className, type, ...props }, ref) => {
+    return (
+        <input
+            type={type}
+            className={cn(
+                "flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-1 text-base shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50 md:text-sm",
+                className,
+            )}
+            ref={ref}
+            {...props}
+        />
+    )
+})
+
+Input.displayName = "Input"
