+++ client-app/src/ui-kit/separator.component.tsx
@@ -1,0 +1,21 @@
+import { Root } from "@radix-ui/react-separator"
+import { cn } from "@/helpers/concat-class-names.helper"
+import { forwardRef, ElementRef, ComponentPropsWithoutRef } from "react"
+
+export const Separator = forwardRef<ElementRef<typeof Root>, ComponentPropsWithoutRef<typeof Root>>(
+    ({ className, orientation = "horizontal", decorative = true, ...props }, ref) => (
+        <Root
+            ref={ref}
+            decorative={decorative}
+            orientation={orientation}
+            className={cn(
+                "shrink-0 bg-border",
+                orientation === "horizontal" ? "h-[1px] w-full" : "h-full w-[1px]",
+                className,
+            )}
+            {...props}
+        />
+    ),
+)
+
+Separator.displayName = Root.displayName
