+++ client-app/src/ui-kit/skeleton.component.tsx
@@ -1,0 +1,5 @@
+import { cn } from "@/helpers/concat-class-names.helper"
+
+export function Skeleton({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
+    return <div className={cn("animate-pulse rounded-md bg-primary/10", className)} {...props} />
+}
